+++ actionpack/lib/abstract_controller/callbacks.rb
@@ -8,7 +8,7 @@ module AbstractController
     include ActiveSupport::Callbacks
 
     included do
-      define_callbacks :process_action, :terminator => "response_body"
+      define_callbacks :process_action, :terminator => "response_body", :skip_after_callbacks_if_terminated => true
     end
 
     # Override AbstractController::Base's process_action to run the
@@ -21,11 +21,9 @@ module AbstractController
 
     module ClassMethods
       # If :only or :except are used, convert the options into the
-      # primitive form (:per_key) used by ActiveSupport::Callbacks.
+      # :unless and :if options of ActiveSupport::Callbacks.
       # The basic idea is that :only => :index gets converted to
-      # :if => proc {|c| c.action_name == "index" }, but that the
-      # proc is only evaluated once per action for the lifetime of
-      # a Rails process.
+      # :if => proc {|c| c.action_name == "index" }.
       #
       # ==== Options
       # * <tt>only</tt>   - The callback should be run only for this action
@@ -33,11 +31,11 @@ module AbstractController
       def _normalize_callback_options(options)
         if only = options[:only]
           only = Array(only).map {|o| "action_name == '#{o}'"}.join(" || ")
-          options[:per_key] = {:if => only}
+          options[:if] = Array(options[:if]) << only
         end
         if except = options[:except]
           except = Array(except).map {|e| "action_name == '#{e}'"}.join(" || ")
-          options[:per_key] = {:unless => except}
+          options[:unless] = Array(options[:unless]) << except
         end
       end
 
@@ -167,7 +165,6 @@ module AbstractController
           # for details on the allowed parameters.
           def #{filter}_filter(*names, &blk)                                                 # def before_filter(*names, &blk)
             _insert_callbacks(names, blk) do |name, options|                                 #   _insert_callbacks(names, blk) do |name, options|
-              options[:if] = (Array(options[:if]) << "!halted") if #{filter == :after}       #     options[:if] = (Array(options[:if]) << "!halted") if false
               set_callback(:process_action, :#{filter}, name, options)                       #     set_callback(:process_action, :before, name, options)
             end                                                                              #   end
           end                                                                                # end
@@ -176,7 +173,6 @@ module AbstractController
           # for details on the allowed parameters.
           def prepend_#{filter}_filter(*names, &blk)                                            # def prepend_before_filter(*names, &blk)
             _insert_callbacks(names, blk) do |name, options|                                    #   _insert_callbacks(names, blk) do |name, options|
-              options[:if] = (Array(options[:if]) << "!halted") if #{filter == :after}          #     options[:if] = (Array(options[:if]) << "!halted") if false
               set_callback(:process_action, :#{filter}, name, options.merge(:prepend => true))  #     set_callback(:process_action, :before, name, options.merge(:prepend => true))
             end                                                                                 #   end
           end                                                                                   # end