+++ actionpack/lib/action_view/helpers/asset_tag_helper.rb
@@ -198,7 +198,7 @@ module ActionView
       include JavascriptTagHelpers
       include StylesheetTagHelpers
       # Returns a link tag that browsers and news readers can use to auto-detect
-      # an RSS or ATOM feed. The +type+ can either be <tt>:rss</tt> (default) or
+      # an RSS or Atom feed. The +type+ can either be <tt>:rss</tt> (default) or
       # <tt>:atom</tt>. Control the link options in url_for format using the
       # +url_options+. You can modify the LINK tag itself in +tag_options+.
       #
@@ -278,6 +278,13 @@ module ActionView
       end
       alias_method :path_to_image, :image_path # aliased to avoid conflicts with an image_path named route
 
+      # Computes the full URL to an image asset in the public images directory.
+      # This will use +image_path+ internally, so most of their behaviors will be the same.
+      def image_url(source)
+        URI.join(current_host, path_to_image(source)).to_s
+      end
+      alias_method :url_to_image, :image_url # aliased to avoid conflicts with an image_url named route
+
       # Computes the path to a video asset in the public videos directory.
       # Full paths from the document root will be passed through.
       # Used internally by +video_tag+ to build the video path.
@@ -292,6 +299,13 @@ module ActionView
         asset_paths.compute_public_path(source, 'videos')
       end
       alias_method :path_to_video, :video_path # aliased to avoid conflicts with a video_path named route
+
+      # Computes the full URL to a video asset in the public videos directory.
+      # This will use +video_path+ internally, so most of their behaviors will be the same.
+      def video_url(source)
+        URI.join(current_host, path_to_video(source)).to_s
+      end
+      alias_method :url_to_video, :video_url # aliased to avoid conflicts with an video_url named route
 
       # Computes the path to an audio asset in the public audios directory.
       # Full paths from the document root will be passed through.
@@ -308,6 +322,13 @@ module ActionView
       end
       alias_method :path_to_audio, :audio_path # aliased to avoid conflicts with an audio_path named route
 
+      # Computes the full URL to a audio asset in the public audios directory.
+      # This will use +audio_path+ internally, so most of their behaviors will be the same.
+      def audio_url(source)
+        URI.join(current_host, path_to_audio(source)).to_s
+      end
+      alias_method :url_to_audio, :audio_url # aliased to avoid conflicts with an audio_url named route
+
       # Computes the path to a font asset in the public fonts directory.
       # Full paths from the document root will be passed through.
       #
@@ -321,6 +342,13 @@ module ActionView
         asset_paths.compute_public_path(source, 'fonts')
       end
       alias_method :path_to_font, :font_path # aliased to avoid conflicts with an font_path named route
+
+      # Computes the full URL to a font asset in the public fonts directory.
+      # This will use +font_path+ internally, so most of their behaviors will be the same.
+      def font_url(source)
+        URI.join(current_host, path_to_font(source)).to_s
+      end
+      alias_method :url_to_font, :font_url # aliased to avoid conflicts with an font_url named route
 
       # Returns an html image tag for the +source+. The +source+ can be a full
       # path or a file that exists in your public images directory.
@@ -460,8 +488,12 @@ module ActionView
             end
           else
             options[:src] = send("path_to_#{type}", sources.first)
-            tag(type, options)
+            content_tag(type, nil, options)
           end
+        end
+
+        def current_host
+          url_for(:only_path => false)
         end
     end
   end