+++ actionpack/test/dispatch/static_test.rb
@@ -1,3 +1,4 @@
+# encoding: utf-8
 require 'abstract_unit'
 
 module StaticTests
@@ -28,6 +29,10 @@ module StaticTests
     assert_html "/foo/index.html", get("/foo/index.html")
     assert_html "/foo/index.html", get("/foo/")
     assert_html "/foo/index.html", get("/foo")
+  end
+
+  def test_served_static_file_with_non_english_filename
+    assert_html "means hello in Japanese\n", get("/foo/#{Rack::Utils.escape("こんにちは.html")}")
   end
 
   private