+++ activesupport/test/callback_inheritance_test.rb
@@ -9,8 +9,8 @@ class GrandParent
   end
 
   define_callbacks :dispatch
-  set_callback :dispatch, :before, :before1, :before2, :per_key => {:if => proc {|c| c.action_name == "index" || c.action_name == "update" }}
-  set_callback :dispatch, :after, :after1, :after2, :per_key => {:if => proc {|c| c.action_name == "update" || c.action_name == "delete" }}
+  set_callback :dispatch, :before, :before1, :before2, :if => proc {|c| c.action_name == "index" || c.action_name == "update" }
+  set_callback :dispatch, :after, :after1, :after2, :if => proc {|c| c.action_name == "update" || c.action_name == "delete" }
 
   def before1
     @log << "before1"
@@ -37,12 +37,12 @@ class GrandParent
 end
 
 class Parent < GrandParent
-  skip_callback :dispatch, :before, :before2, :per_key => {:unless => proc {|c| c.action_name == "update" }}
-  skip_callback :dispatch, :after, :after2, :per_key => {:unless => proc {|c| c.action_name == "delete" }}
+  skip_callback :dispatch, :before, :before2, :unless => proc {|c| c.action_name == "update" }
+  skip_callback :dispatch, :after, :after2, :unless => proc {|c| c.action_name == "delete" }
 end
 
 class Child < GrandParent
-  skip_callback :dispatch, :before, :before2, :per_key => {:unless => proc {|c| c.action_name == "update" }}, :if => :state_open?
+  skip_callback :dispatch, :before, :before2, :unless => proc {|c| c.action_name == "update" }, :if => :state_open?
 
   def state_open?
     @state == :open
@@ -112,15 +112,15 @@ class BasicCallbacksTest < ActiveSupport::TestCase
     @unknown  = GrandParent.new("unknown").dispatch
   end
 
-  def test_basic_per_key1
+  def test_basic_conditional_callback1
     assert_equal %w(before1 before2 index), @index.log
   end
 
-  def test_basic_per_key2
+  def test_basic_conditional_callback2
     assert_equal %w(before1 before2 update after2 after1), @update.log
   end
 
-  def test_basic_per_key3
+  def test_basic_conditional_callback3
     assert_equal %w(delete after2 after1), @delete.log
   end
 end